app/src/test/java/com/cevin/submission3/ui/tvshow/TvShowViewModelTest.kt
package com.cevin.submission3.ui.tvshow

import androidx.arch.core.executor.testing.InstantTaskExecutorRule
import androidx.lifecycle.MutableLiveData
import androidx.lifecycle.Observer
import com.cevin.submission3.data.TvShow
import com.cevin.submission3.data.source.TheMovieRepository
import com.cevin.submission3.utils.DataDummy
import org.junit.Assert.assertEquals
import org.junit.Assert.assertNotNull
import org.junit.Before
import org.junit.Rule
import org.junit.Test
import org.junit.runner.RunWith
import org.mockito.Mock
import org.mockito.Mockito
import org.mockito.junit.MockitoJUnitRunner
import org.mockito.kotlin.verify

@RunWith(MockitoJUnitRunner::class)
class TvShowViewModelTest {
    private lateinit var tvShowViewModel: TvShowViewModel

    @get:Rule
    var instantTaskExecutorRule = InstantTaskExecutorRule()

    @Mock
    private lateinit var observer: Observer<List<TvShow>>

    @Mock
    private lateinit var theMovieRepository: TheMovieRepository

    @Before
    fun setUp() {
        tvShowViewModel = TvShowViewModel(theMovieRepository)
    }

    @Test
    fun getListTvShow() {
        val dummyTv = DataDummy.generateDummyTvShow()
        val tvShow = MutableLiveData<List<TvShow>>()
        tvShow.value = dummyTv

        Mockito.`when`(theMovieRepository.getAllTvShow(1)).thenReturn(tvShow)
        val listTv = tvShowViewModel.getListTvShow()
        verify(theMovieRepository).getAllTvShow(1)

        listTv.observeForever(observer)
        verify(observer).onChanged(dummyTv)

        assertNotNull(listTv)
        assertEquals(20, listTv.value?.size ?: 0)
    }
}